Reviewed by:	Jordan Submitted by:	Rich
Make conditionals on BIOSWAIT consistent with usage in io.c.

If you had 'BOOTWAIT=0' in your /etc/make.conf then biosboot wouldn't
compile.  It was '#if' in io.c and '#ifdef' in probe_keyboard.c so I
changed the latter to '#if'.

Even if BOOTWAIT is undefined then '#if BOOTWAIT' becomes
'#if 0' so it should compile either way with this change.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@6419 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f