Remove obsolete comments about my-name and my-network.
Have netstart directly source sysconfig so that it stands on its own.
Do not source netstart in rc, run it with sh.

Rework the dangerous /etc/hostname.* so that it uses a variable with
a list of interfaces and a variable for each interface in that list.
The files /etc/hostname.* become obsolete with this change, the
information is now stored in /etc/sysconfig.

Source any /etc/start_if.${ifn} files so they can export things to
netstart (plans for future enhancements in this area.)

Obsolete /etc/defaultrouter, this is now down with $defaultrouter from
sysconfig.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@7750 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f