Do not allow the caller to specify the same path for the special device file and the mount point.  This prevents the "unexpected recursive lock" panic from happening.
This is a temporary fix.  A kernel fix would be much much more ugly than
this, and still wouldn't be the "right" way to fix it.  After some
of Terry's file system rework is installed, it will be possible to
properly fix this problem in a clean manner.  Until then,
this change should prevent use from getting a problem report
on this every month or so (and I just noticed that someone in
one of the freebsd news groups was complaining about this problem, too).


git-svn-id: svn+ssh://svn.freebsd.org/base/head@14445 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f