Make some changes to the way the label editor reads script variables so you don't need to re-enter it for each and every filesystem. Heads up! This change is incompatible with the previous scripting format, so those folks (all 2 of you) using config files should take a look at the changes to the sample install.cfg file for the diskLabelEditor's new calling syntax.
Finally write a man page for this thing, documenting all of the above
and more.  I can't drive a stake through this thing's heart without
properly documenting it first, so please consider this step #1 in that
process (to be honest, sysinstall will also live on for some time in
the 2.2. branch since it's unlikely that the new install tools will ever
make it over there - they're strictly 3.0 material).


git-svn-id: svn+ssh://svn.freebsd.org/base/head@28075 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f