o  make SHARED=symlinks a caller defined instead of a callee defined    property. This fixes the includes target when DESTDIR is empty. o  Do not make build-tools for f771 when NO_FORTRAN is defined. o  Add new build stage. See below. o  Change banners so that staging information is displayed.
The addition of the build-tools target broke the upgrade path because
we couldn't make use of previously built tools that were made for
compatibility reasons. Doing so would also result in the cross-compiler
being used and that is exactly what had to be avoided.

This is solved by designating the bootstrap-tools stage for building
anything that is needed for compatibility only and to create a new
stage (started after the build-tools stage) that handles cross-tools
building. We now have the following stages:
1. bootstrap-tools (for compatibility issues only)
2. build-tools
3. cross-tools (what it says)
4. world
5. install

Stages 1-4 (inclusive) are handled by buildworld.
Stage 5 is handled by installworld.

Any more stages and I'll join Nik in his quest for the
holy grail^W^Wworld :-)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@54875 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f