New machine independant /dev/null and /dev/zero driver. This device is severely stripped down compared with its predecessor, and is measurably a _lot_ faster.
Many thanks to Jeroen van Gelderen for lots of good ideas.

There is still a problem with this; it is written as a mudule, and as
such is theoretically unloadable. However, there is no refcounting done
as I would prefer to do that a'la device_busy(9), rather than some
"home-rolled" scheme. The point is pretty moot, as /dev/null is
effectively compulsory.

Reviewed by:	dfr


git-svn-id: svn+ssh://svn.freebsd.org/base/head@62052 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f