Do not panic on an uninitialized VOP_xxx() call.  This was meant as a sanity check, but it is too easy to run into, eg: making an ACL syscall when no filesystems have the ACL implementation enabled.
The original reason for the panic was that the VOP_ vector had not been
assigned and therefor could not be passed down the stack.. and there
was no point passing it down since nothing implemented it anyway.
vop_defaultop entries could not pass it on because it had a zero (unknown)
vector that was indistinguishable from another unknown VOP vector.

Anyway, we can do something reasonable in this case, we shouldn't need
to panic here as there is a reasonable recovery option (return EOPNOTSUPP
and dont pass it down the stack).

Requested by: rwatson


git-svn-id: svn+ssh://svn.freebsd.org/base/head@65527 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f