Add __va_copy and make it always visible, in spite of the __ISO_C_VISIBLE setting.  Make va_copy be an alias if __ISO_C_VISIBLE >= 1999.
Why?  more than a few ports have an autoconf that looks for __va_copy
because it is available on glibc.  It is critical that we use it if
at all possible on amd64.  It generally isn't a problem for i386 and its
ilk because autoconf driven code tends to fall back to an assignment.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@121450 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f