Introduce the SO_BINTIME option which takes a high-resolution timestamp at packet arrival.
For benchmarking purposes SO_BINTIME is preferable to SO_TIMEVAL
since it has higher resolution and lower overhead.  Simultaneous
use of the two options is possible and they will return consistent
timestamps.

This introduces an extra test and a function call for SO_TIMEVAL, but I have
not been able to measure that.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@125264 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f