Recognize hardlinks when reading cpio files.
This doesn't yet address the issue of selective restore
of hardlinked files.  With cpio format, it's possible to correctly
restore any linked file; the API doesn't yet fully support this.
(There's no way for the library to inform a client whether or not
there's a file body associated with this entry.  The assumption
right now is that "hardlink" entries have no file body.)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@126705 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f