Remove the old ATA_*LOCK_CH macros that used atomic ops and use mutexes instead. This closes the last (known) race issues in ATA which should fix the various hangs etc seen on heavy loaded systems.
Change from using timeout functions to using callout functions in
the timeout code. This together with above closes the race that could
happen if timeout and device interrupt occured simultaniously.

Also fix the possible recursion in ata_reinit() on very dodgy
devices that could take us down in the probe.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@135819 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f