Turns out that revision 1.52 was a bad idea.  It broke the long standing ability to list a non-existant device in /etc/ttys to keep it from dying.  This is a documented feature of init(8):      The init utility can also be used to keep arbitrary daemons running,      automatically restarting them if they die.  In this case, the first field      in the ttys(5) file must not reference the path to a configured device      node and will be passed to the daemon as the final argument on its com-      mand line.  This is similar to the facility offered in the AT&T System V      UNIX /etc/inittab.
So rather than fix the man page to 'break' this feature, back out the change.

At the time this change was made, people felt that the spamage from
getty was annoying on headless consoles.  Andrew Gallatin noted:
> Most of my machines are headless without video cards and use a serial
> console.  With devfs this means that /dev/ttyv[1-N] do not exist and
> getty bitches like this:
>
> Sep 26 11:00:11 monet getty[543]: open /dev/ttyv1: No such file or directory

and we went off and applied this hack rather than fixing getty to
sleep forever when it gets an unknown device, as was Andrew's other
suggestion.  Since it breaks things, I'm off to do that instead.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@135868 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f