There were two calls to cardbus_do_cis when cardbus_do_cis changed, yet I only changed one of them.  So when we loaded drivers, we'd fail to allocate resources correct.
This pointed out that we were doing the wrong thing when we failed to
attach a child.  We released all the resources and almost deleted the
child.  Instead, we should keep the resources allocated so when/if a
driver is loaded, we can go w/o having to allocate them.  We use
pci_cfg_save/restore to restore the BARs with these resources.

This seems to fix the problems that we were seeing that I thought
might have magically gone away in the last revision of cardbus.c (but
really didn't).

Noticed by: avatar (nicely done!)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@142736 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f