Attempt to complete the userspace integration of POSIX.1e extended ACLs. This includes adding support for ACLs into cp(1) and mv(1) userspace utilities.
For mv(1), if _PC_ACL_EXTENDED is in effect for the source AND destination
operands, the destination file's ACLs shall reflect the source.

For cp(1), if _PC_ACL_EXTENDED is in effect for both source and destination
operands, and -p has been specified, the ACLs from the source shall be
preserved on the destination.

MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@149790 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f