Nuke duplicates found via fuzzy logic: - typos - different spelling, punctuation, whitespace - phonetically similar names - words rearranged ("was once" vs "once was" etc)
If a limerick appeared as a single one and as part of a
double or triple, the singleton was removed.

With a little help from: sort limerick|uniq -d
This still turns up 20 lines being repeated, but the respective
limericks are sufficiently unique to leave them in (i.e. most differ
in at least two lines).

Nuke spaces in front of colons while I'm here.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@150699 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f