Back out most of 1.84.  It was unwise to force debug kernels to always be installed.  It should have been optional to install a non-debug one, just like it was formerly optional to install a debug one.  In order to do that, most of 1.84 had to go.
Instead, make installing the debug kernel the default, but create a
new option INSTALL_NODEBUG for those people that have small /
partitions and good source control habits.

This preserves the behavior of 1.84 while allowing it to be overriden
for people (like me) that do not have the time to upgrade to get a
bigger / and also don't have time for stupid makefile tricks when
upgrading their older system, but still want a kernel.debug around if
things go south.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@151636 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f