Change the semantics of brk_max to dynamically deal with data segment bounds. [1]
Modify logic for utilizing the data segment, such that it is possible to
create huge allocations there.

Shrink the data segment when deallocating a chunk, if it is at the end of
the data segment.

Rename chunk_size to csize in huge_malloc(), in order to avoid masking a
static variable of the same name. [1]

Reported by:	Paul Allen <nospam@ugcs.caltech.edu>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@158062 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f