Remove dependency on ispfw and firmware as modules. Either they're there early and the ispfw sets have registered themselves, or they're not.
The module dependency stuff isn't quite what we want
anyway. If the user doesn't want the load placed on
system memory by loading the firmware, they don't
specify it to be loaded (either by being linked in
or via being a module to be loaded and then hooked
in with firmware(9)). It doesn't then make sense to
then override what they want by pulling it in anyway.

This might be able to work if we were able to pull in
just exactly what we needed for the card we have- but
that's an optimization left for the future.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@165061 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f