sysctl_rman: report shared resources to devinfo
shared uses of a resource are recorded on a sub-list hanging off
a main resource object on a main resource list;
without this change a shared resource (e.g. irq) is reported only
once by devinfo -r/-u;
with this change the resource is reported for each driver that
allocates it (which is even more than what vmstat -i -a reports).

Approved by:	jhb (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@192379 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f