Fix two bugs in r218670.
Hold the vnode around the region where object lock is dropped, until
vnode lock is acquired.

Do not drop the vnode reference for a case when the object was
deallocated during unlock. Note that in this case, VV_TEXT is cleared
by vnode_pager_dealloc().

Reported and tested by:	pho
Reviewed by:	alc
MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@220977 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f