Change the default CABQ time to be 70% of the beacon interval, rather than the whole beacon interval.
The reference driver and Linux ath9k both choose 80% of the
beacon interval and they do it in the driver rather than
the HAL (Ath reference) or ath9k_hw (ath9k.)

This quietens stuck beacon conditions on my AR9220/AR9280
based NICs when a lot of burst broadcast/multicast traffic
is going on. It doesn't seem to annoy the earlier MACs as
much as the AR9280 and later one.

Obtained from:	Linux ath9k, Atheros


git-svn-id: svn+ssh://svn.freebsd.org/base/head@225820 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f