kern cons: introduce infrastructure for console grabbing by kernel
At the moment grab and ungrab methods of all console drivers are no-ops.

Current intended meaning of the calls is that the kernel takes control of
console input.  In the future the semantics may be extended to mean that
the calling thread takes full ownership of the console (e.g. console
output from other threads could be suspended).

Inspired by:	bde
MFC after:	2 months


git-svn-id: svn+ssh://svn.freebsd.org/base/head@228631 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f