Hide kernel option ROUTETABLES evaluations in the implementation rather than the header file.  With this also move RT_MAXFIBS and RT_NUMFIBS into the implemantion to avoid further usage in other code. rt_numfibs is all that should be needed.
This allows users to change the number of FIBs from 1..RT_MAXFIBS(16)
dynamically using the tunable without the need to change the kernel
config for the maximum anymore.  This means that thet multi-FIB
feature is now fully available with GENERIC kernels.
The kernel option ROUTETABLES can still be used to set the default
numbers of FIBs in absence of the tunable.

Ok.ed by:	julian, hrs, melifaro
MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@233113 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f