Add isnan() and isinf() to the global namespace in libstdc++'s <cmath>.
The standard (n3242, section 17.6.1.1, paragraph 4) says that, because these are
declared as macros in the C specification (even though they are
implemented as functions in the C++ library) they should be in the global
namespace.

A surprising number of configure checks rely on this.  It was broken by recent
cleanups to math.h.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@253563 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f