Fix broken SMP startup on Armada XP after r265694
During Armada's platform_mp_start_ap(), mptramp code
is being copied to the specific physical location (0xffff0000).
Before r265694 the address to which the code should be copied
was equal to the address of mpentry routine that followed the
mptramp in locore.S. Now the mptramp end address should be
exported and used as a copy limit.

Reviewed by: gber


git-svn-id: svn+ssh://svn.freebsd.org/base/head@267129 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f