Fix alignment directives in arm asm code after clang 3.5 import.
The ancient gas we've been using interprets .align 0 as align to the
minimum required alignment for the current section.  Clang's integrated
assembler interprets it as align to a byte boundary.  Fortunately both
assemblers interpret a non-zero value as align to 2^N so just make sure
we have appropriate non-zero values everywhere.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@276596 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f