Teach bits of libc about Thumb. This adds the if-then instructions needed to handle the ARM conditional execution.
While here fix a bug found by this in the hard-float code, cc is the
opposite of cs. The former is used for 'less than' in floating-point code
and is executed when the C (carry) bit is clear, the latter is used when
greater than, equal, or unordered, and is executed when the C bit is set.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@282816 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f