Fix decoding of UTF-7 when a base64 encoded chunk appears at the end of the input buffer.
_citrus_UTF7_mbtoutf16 stored the decoder state at the beginning so it
could restore this state on an incomplete character such that the next
call would restart the decoding.  The problem was that "-" (end of base64
mode) at the end of a string was also treated as an incomplete character
but was also removed from the state buffer.  So the initial state would be
restored (with base64 mode) and the next call would no longer see the "-"
so it continued in base64 mode.

This state saving/restoring isn't needed here.  It's already handled
elsewhere (citrus_iconv_std.c:_citrus_iconv_std_iconv_convert) so just
remove it.

Also initialise *nresult.

PR:		200398
Tested by:	delphij
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@283406 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f