Correct the polarity of the PPS assert and clear events with respect to the electrical signals on the serial port.  Virtually all devices which output a PPS signal generate a brief higher-voltage pulse, the leading edge of which is the on-time point.
Both DCD and CTS are active-low signals on the wire, meaning the assertion
of their status bits in the modem status register corresponds to the lower
voltage level on the wire.  So when the status bit transitions to not-set,
create a PPS assert event; when the status bit transitions to set, create a
PPS clear event.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@286595 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f