Add an initial driver for the AR9170 series draft-11n hardware from Atheros.
Thanks to OpenBSD for providing a driver based on the original
Atheros open source driver circa 2008.  This uses the early, pre-carl9170
atheros provided firmware.

It only supports 11bg at the moment.  I've not tested it with 11a
(and so the TX rate control logic may be slightly wrong!) so if
you do have the dual-band version of this hardware please do let me know.

Tested:

* AR9170, TP-Link WN821N 2GHz.

TODO:

* Hook this up to a non-module build.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@288253 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f