Abstract out the ampdu TX pps initialisation code so it can be reused in the superg fast-frames code.
This harks back to an earlier commit (r280349) where I found that
initialising the pps code with ticks=0 would cause hilariously bad
hz ticks wraparound failures, leading to never actually aggregating
traffic.  This is still true for the superg path and so I have to
do the same thing there.

This is a big no-op; a subsequent commit will flip this on so it
works with the fast-frames transmit path.

Tested:

* AR9170, otus(4) - STA mode, 11bg operation
* AR9331, AP mode


git-svn-id: svn+ssh://svn.freebsd.org/base/head@288315 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f