The latest version of lex requires the latest m4 to build, add a dependency when running the build-tools stage.
The requirement is due to the -P flag used when running m4 from usr.bin/lex
Makefile to generate skel.c.  With the old m4 that fails and the failure is
ignored, resulting in an empty(-ish) skel.c, which leads to later build
failures when the misconfigured new lex tool is run.

This enables building -current (and 10-stable after MFC) on a stable-8
system again.

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@288829 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f