Fix the %b flags string for ddb.  All bits above the 5th (TF_OPENED_CONS) were broken in r188147 by adding TF_OPENED_CONS without updating the string.  It was especially confusing to display OPENED_CONS as GONE and BYPASS as ZOMBIE.  2 flags at the end were not updated in r188487.
Don't print an extra 0x prefix for %p in a ddb command.  In the rest
of the kernel there are more than 6000 lines with %p and only about
40 with this bug.

Print a non-extra 0x prefix for %b in a ddb command.  In the rest
of the kernel, there are approx. 180 lines with %b and 2/3 of them
have this bug.

Submitted by:	bde
MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@294732 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f