Add a tool to decode ioctl commands.
One or more ioctl command values can be passed as arguments on the command
line.  For each value, the command is broken down into it's components
(direction, group, number, and length).  In addition, if a command has a
known name it is output via sysdecode_ioctlname().

Reviewed by:	kib, emaste, avg
Approved by:	re (gjb)
Differential Revision:	https://reviews.freebsd.org/D6851


git-svn-id: svn+ssh://svn.freebsd.org/base/head@301931 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f