Avoid truncating the return value of DTrace predicates.
Predicates are DIF objects whose return value is compared with zero to
determine whether the corresponding probe body is to be executed. The return
value itself is the contents of a 64-bit DIF register, but it was being
truncated to an int before the comparison. This meant that a predicate such
as /0x100000000/ would evaluate to false.

Reported by:	rwatson
MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@302507 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f