sed(1): Revert r303047 "cleanup" and therefore r303572.
While big, the change was meant to have no effect on behavior and instead
so far we have found two regressions: one in the etcupdate tests and
another one in the games/openttd port[1].

Revert to a known working state. We will likely have to split the patch in
functional parts before bringing back the changes.

PR:		195929
Reported by:	danfe, madpilot [1]


git-svn-id: svn+ssh://svn.freebsd.org/base/head@303662 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f