Correct the type of db_cmd_loop_done.
On big endian hardware that uses 1 byte bool a type mismatch of bool vs int will
cause the least signifcant byte of db_cmd_loop_done to be set, but the MSB to be
read, and read as 0.  This causes ddb to stay in an infinite loop.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@305636 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f