xen/intr: add reference counts to event channels
Add a reference count to xenisrc. This is required for implementation of
unmap-notifications in the grant table userspace device (gntdev). We need to
hold a reference to the event channel port, in case the user deallocates the
port before we send the notification.

Submitted by:		jaggi
Reviewed by:		royger
Differential review:	https://reviews.freebsd.org/D7429


git-svn-id: svn+ssh://svn.freebsd.org/base/head@308127 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f