indent(1): Don't ignore newlines after comments that follow braces.
indent.c has a special loop that stores tokens from between an if () and
the next statement into a buffer. The loop ignored all newlines, but that
resulted in not calling dump_line() when it was needed to produce the
final line of the buffered up comment.

Taken from:	 Piotr Stefaniak


git-svn-id: svn+ssh://svn.freebsd.org/base/head@309342 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f