Revert r253678, r253661: Fix a segfault in ctfmerge(1) due to a bug in GCC.
The change was correct and the bug real, but upstream didn't adopt it
and we want to remain in sync. When/if upstream does something about it
we can bring their version.

The bug in question was fixed in GCC 4.9 which is now the default in
FreeBSD's ports. Our native gcc-4.2, which is still in use in some Tier-2
platforms also has a workaround so no end-user should be harmed by the
revert.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@309490 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f