Add rfdi opcode to binutils
The rfdi instruction is part of the e500mc and derivative cores.  It came into
binutils in a GPLv3 patch, along with the rest of the e500mc instruction set.
Currently only rfdi is planned to be used, so rather than attempt to backport
the full patch, take a surgical route and add instructions as needed.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@313035 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f