cxgbe/iw_cxgbe: Implement sq/rq drain operation.
ULPs can set a qp's state to ERROR and then post a work request on the
sq and/or rq.  When the reply for that work request comes back it is
guaranteed that all previous work requests posted on that queue have
been drained.

Obtained from:	Chelsio Communications
MFC after:	3 days
Sponsored by:	Chelsio Communications


git-svn-id: svn+ssh://svn.freebsd.org/base/head@314578 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f