Remove tangled isp_mbox_continue() mechanism.
It was implemented to reduce context switches when uploading firmware to
card's RAM.  But this mechanism is not used last 10 years since all mbox
operations are now polled, and it was never used for cards produced in
last 15 years.  Newer cards can use DMA to upload firmware.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@315273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f