Fix stack tracing in dtrace for powerpc
The current method only sort of works, and usually doesn't work reliably.
Also, on Book-E the return address from DEBUG exceptions is not the sentinel
addresses, so it won't exit the loop correctly.

Fix this by better handling trap frames during unwinding, and using the
common trap handler for debug traps, as the code in that segment is
identical between the two.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@318167 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f