Remove groff from base
All manpages in base are now compatible with mandoc(1), all roff documentation
will be relocated in the doc tree. man(1) can now use groff from the ports tree
if it needs.

Also remove checknr(1) and colcrt(1) which are only useful with groff.

Approved by:	(no objections on the mailing lists)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@319664 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f