packages: Allow stageworld/stagekernel to run with make jobs.
The -B was originally added in projects/release-pkg r289381 as a copy
of what 'make world' did at the time.  The -B was removed from
the 'installworld' call in 'world' in r303844 though.  The staging
of files is safe to run in parallel.

MFC after:	2 weeks
Sponsored by:	Dell EMC Isilon


git-svn-id: svn+ssh://svn.freebsd.org/base/head@320280 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f