sh: Ignore error when cd writes the directory actually switched to.
If CDPATH is used non-trivially or the operand is "-", cd writes the
directory actually switched to. (We currently do this only in interactive
shells, but POSIX requires this in non-interactive shells as well.)

As mentioned in Austin group bug #1045, cd shall not return an error while
leaving the current directory changed. Therefore, ignore any write error.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@320340 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f