sx: unbreak debug after r326107
An assertion was modified to use the found value, but it was not updated to
handle a race where blocked threads appear after the entrance to the func.

Move the assertion down to the area protected with sleepq lock where the
lock is read anyway. This does not affect coverage of the assertion and
is consistent with what rw locks are doing.

Reported by:	Shawn Webb


git-svn-id: svn+ssh://svn.freebsd.org/base/head@326112 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f