amd-vi: set iommu msi configuration using pci_enable_msi method
This is better than directly changing PCI configuration space of the
device because it makes the PCI bus aware of the configuration.
Also, the change allows to drop a bunch of code that duplicated
pci_enable_msi() functionality.

I wonder if it's possible to further simplify the code by using
pci_alloc_msi().


git-svn-id: svn+ssh://svn.freebsd.org/base/head@326539 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f