{ext2|ufs}_readdir: Set limit on valid ncookies values.
Sanitize the values that will be assigned to ncookies so that we ensure
they are sane and we can handle them.

Let ncookies signed as it was before r328346. The valid range is such
that unsigned values are not required and we are not able to avoid at
least one cast anyways.

Hinted by:	bde


git-svn-id: svn+ssh://svn.freebsd.org/base/head@328479 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f