{ext2|ufs}_readdir: Avoid setting negative ncookies.
ncookies cannot be negative or the allocator will fail. This should only
happen if a caller is very broken but we can still try to survive the
event.

We should probably also verify for uio_resid > MAXPHYS but in that case
it is not clear that just clipping the ncookies value is an adequate
response.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@328957 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f