stand/lua: Call menu_entries if it's a function
If we've fetched menu.entries and it turns out it's a function, call it to
get the actual menu entries.

This will be used to swap multi-/single- user boot options if we're booting
single user by default (boot_single="YES" in loader.conf(5)). It can also be
used fairly easily for other non-standard situations.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@329586 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f