stand/lua: Add and use drawer.menu_name_handlers
Pull out specialized naming behavior into drawer.menu_name_handlers. This is
currently only needed for the carousel entry, where naming is based on the
current choice and the menu item purposefully does not store this state.

The setup was designed so that every type can have a special name handler,
and the default action is to simply use the result of entry.name().


git-svn-id: svn+ssh://svn.freebsd.org/base/head@329621 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f