lualoader: Pull menu redrawing specifics out of menu.process
In general, every menu redraw is going to require a screen clear and cursor
reset. Each redraw also has the potential to invalidate the alias table, so
we move the alias table being used out into a module variable. This allows
third party consumers to also inspect or update the alias table if they need
to.

While here, stop searching the alias table once we've found a match.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@329947 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f