lualoader: Remove debug function
Our module bits ended up more stable than I anticipated, so this turns out
to be no longer useful.

If things like this need to come back, we should do it in a separate 'debug'
module to serve as a collection of debugging aides. As a rule, this 'debug'
module would *not* be allowed as a requirement of any other modules in-tree.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@330100 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f