loader efifb: implement uga_autoresize as a call to text_autoresize
UGA does not have the same kind of mode enumeration that GOP does. Implement
it instead as a call to text_autoresize so that firmwares with only UGA
present still get some kind of autoresizing behavior.

While here, rename a typo'd "gop" to "uga", although it will remain unused
for the time being.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@331547 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f